Random Forest Refactored Random Forest Code and added Example.
