+++ algorithms/linfa-ensemble/src/random_forest/mod.rs
@@ -1,5 +1,3 @@
 mod algorithm;
-mod hyperparams;
 
 pub use algorithm::*;
-pub use hyperparams::*;
